+++ src/Filter.h
@@ -60,8 +60,7 @@ public:
     virtual bool accepts(void *data) = 0;
     virtual void *indexFilter(const char *columnname __attribute__ ((__unused__))) { return 0; }
     virtual void findIntLimits(const char *columnname __attribute__ ((__unused__)), int *lower __attribute__ ((__unused__)), int *upper __attribute__ ((__unused__))) {}
-    virtual bool optimizeBitmask(const char *columnname __attribute__ ((__unused__)), uint32_t *mask __attribute__ ((__unused__))) { return false; }
+    virtual bool optimizeBitmask(const char *columnname __attribute__ ((__unused__)), uint64_t *mask __attribute__ ((__unused__))) { return false; }
 };
 
 #endif // Filter_h
-
