+++ src/HostSpecialIntColumn.h
@@ -40,8 +40,7 @@ class HostSpecialIntColumn : public IntColumn
 public:
     HostSpecialIntColumn(string name, string description, int hsic_type, int indirect)
         : IntColumn(name, description, indirect) , _type(hsic_type) {}
-    int32_t getValue(void *data, Query *);
+    int64_t getValue(void *data, Query *);
 };
 
 #endif // HostSpecialIntColumn_h
-
