+++ src/IntColumn.h
@@ -38,7 +38,7 @@ class IntColumn : public Column
 public:
     IntColumn(string name, string description, int indirect_offset)
         : Column(name, description, indirect_offset) {}
-    virtual int32_t getValue(void *data, Query *) = 0;
+    virtual int64_t getValue(void *data, Query *) = 0;
     void output(void *, Query *);
     int type() { return COLTYPE_INT; }
     string valueAsString(void *data, Query *);
@@ -47,4 +47,3 @@ public:
 };
 
 #endif // IntColumn_h
-
