+++ src/IntColumnFilter.h
@@ -42,12 +42,11 @@ class IntColumnFilter : public Filter
 
 public:
     IntColumnFilter(IntColumn *column, int opid, char *value);
-    virtual int32_t convertRefValue(); // see TimeColumnFilter
+    virtual int64_t convertRefValue(); // see TimeColumnFilter
     bool accepts(void *data);
     void findIntLimits(const char *columnname, int *lower, int *upper);
-    bool optimizeBitmask(const char *columnname, uint32_t *mask);
+    bool optimizeBitmask(const char *columnname, uint64_t *mask);
 };
 
 
 #endif // IntColumnFilter_h
-
