+++ src/Logfile.cc
@@ -181,7 +181,7 @@ long Logfile::freeMessages(unsigned logclasses)
     return freed;
 }
 
-inline bool Logfile::processLogLine(uint32_t lineno, unsigned logclasses)
+inline bool Logfile::processLogLine(uint64_t lineno, unsigned logclasses)
 {
     LogEntry *entry = new LogEntry(lineno, _linebuffer);
     // ignored invalid lines
