+++ src/Logfile.h
@@ -49,7 +49,7 @@ private:
     bool       _watch;         // true only for current logfile
     ino_t      _inode;         // needed to detect switching
     fpos_t     _read_pos;      // read until this position
-    uint32_t   _lineno;        // read until this line
+    uint64_t   _lineno;        // read until this line
 
     logfile_entries_t  _entries;
     char       _linebuffer[MAX_LOGLINE];
@@ -80,7 +80,7 @@ public:
 private:
     void loadRange(FILE *file, unsigned missing_types, LogCache *,
                    time_t since, time_t until, unsigned logclasses);
-    bool processLogLine(uint32_t, unsigned);
+    bool processLogLine(uint64_t, unsigned);
     uint64_t makeKey(time_t, unsigned);
 };
