+++ src/OffsetIntColumn.cc
@@ -26,15 +26,14 @@
 #include <stdint.h>
 #include "OffsetIntColumn.h"
 
-int32_t OffsetIntColumn::getValue(void *data, Query *)
+int64_t OffsetIntColumn::getValue(void *data, Query *)
 {
     if (!data)
         return 0;
 
     char *p = (char *)shiftPointer(data);
     if (p)
-        return *(int *)(p + _offset);
+        return *(int64_t *)(p + _offset);
     else
         return 0;
 }
-
