+++ src/ServiceSpecialIntColumn.h
@@ -38,8 +38,7 @@ class ServiceSpecialIntColumn : public IntColumn
 public:
     ServiceSpecialIntColumn(string name, string description, int ssic_type, int indirect)
         : IntColumn(name, description, indirect) , _type(ssic_type) {}
-    int32_t getValue(void *data, Query *);
+    int64_t getValue(void *data, Query *);
 };
 
 #endif // ServiceSpecialIntColumn_h
-
